refactor : vaildate util 함수 폴더구조 변경
