+++ hooks/src/lib/util/validateCVCNumber/validateCVCNumber.ts
@@ -1,5 +1,5 @@
-import { CVC_LENGTH, ERROR_MESSAGE } from "../constants";
-import { isNumeric } from "./index";
+import { CVC_LENGTH, ERROR_MESSAGE } from "../../constants";
+import { isNumeric } from "../index";
 export default function validateCVCNumber(CVCNumber: string) {
   const trimCVCNumber = CVCNumber.trim();
   if (!isNumeric(trimCVCNumber)) return ERROR_MESSAGE.NOT_NUMERIC;
