+++ hooks/src/lib/util/validatePasswordNumber/validatePasswordNumber.ts
@@ -1,5 +1,5 @@
-import { PASSWORD_LENGTH, ERROR_MESSAGE } from "../constants";
-import { isNumeric } from "./index";
+import { PASSWORD_LENGTH, ERROR_MESSAGE } from "../../constants";
+import { isNumeric } from "../index";
 export default function validatePasswordNumber(passwordNumber: string) {
   const trimPasswordNumber = passwordNumber.trim();
   if (!isNumeric(trimPasswordNumber)) return ERROR_MESSAGE.NOT_NUMERIC;
